Added basic framework for operations
Added the basic framework for operations, having a class for each one
and function "operation" that operates on an array based on the type of
operation. Made array of operations in kenken class. Also made minor fix
to the cellGroup class so now currentSize gets incremented when more
cells are added.